08c Slideshow Python and Photoshop/requirements.txt
comtypes==1.4.4
photoshop-python-api==0.22.4
pillow==10.3.0
pypiwin32==223
pywin32==306
